# jit: Implement register allocation