--- a/src/jit.h
+++ b/src/jit.h
@@ -40,8 +40,9 @@ struct jit_state {
 struct host_reg {
     uint8_t reg_idx : 5;   /* index to the host's register file */
     int8_t vm_reg_idx : 6; /* index to the vm register */
-    bool dirty : 1;
-    bool alive : 1;
+    bool dirty : 1; /* whether the context of register has been overridden */
+    bool alive : 1; /* whether the register is no longer used in current basic
+                       block */
 };
 
 struct jit_state *jit_state_init(size_t size);