Implement the register allocator with linear-scan register allocation. The `reg_table` is removed and integrated to the `host_reg` for convenience of accessing the vm register by the `reg_idx`.

The performance comparsion is shown below:

* x86-64

| Metric    | Original       | w/ reg-alloc   | SpeedUp |
| --------- | --------------- | -------------- | ------- |
| dhrystone |  14060 DMIPS   |  15225 DMIPS   | +8.29%  |
| coremark  |   3874 iters/s |   3828 iters/s | -1.19%  |
| aes       | 0.0297 s       | 0.0323 s       | -8.75%  |
| nqueens   |   2.16 s       |   2.05 s       | +5.09%  |
| stream    |   29.6 s       |   27.6 s       | +6.76%  |