@vacantron 2024-04-26
> Enforce consistent naming scheme, in the form `item_action`. E.g., prefer `liveness_calc` instead of `calc_liveness`.

Is it better to use `liveness_reset`, `reg_pick` instead of `reset_liveness`, `pick_reg` also?