@jserv 2024-04-26
> Is it better to use `liveness_reset`, `reg_pick` instead of `reset_liveness`, `pick_reg` also?

Yes, to avoid unintended symbol conflicts, let's stick to the requested naming scheme across compilation units.